requirements.txt
Appium-Python-Client==1.0.2
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
